+++ src/main/java/hu/gov/nav/answerdraft/SourceType.java
@@ -1,0 +1,6 @@
+package hu.gov.nav.answerdraft;
+
+enum SourceType {
+    PUBLIC,
+    APPROVED_KNOWLEDGE
+}